src/requirements.txt
pyspf
dnspython